requirements.txt
Pillow==9.5.0
numpy==1.24.1
IOPaint==1.3.3
ultralytics==8.2.28
opencv_python_headless==4.9.0.80